Requested changes.

+++ server/modules/watchers/watchers.module.ts
@@ -150,11 +150,13 @@ export function initWatchers(models: IModelsParam, broadcast: BroadcastCallback,
 			case 'removed': {
 				const trash = await Messages.trashFindOneById<Pick<IMessage, 'u' | 'rid' | 'file'>>(id, { projection: { u: 1, rid: 1, file: 1 } });
 				const message = trash;
-				const room = trash?.rid ? await Rooms.findOneById<IRoom>(trash.rid, { projection: roomFields }) : undefined;
+				const room = trash?.rid ? await Rooms.trashFindOneById<IRoom>(trash.rid, { projection: roomFields }) : undefined;
 				if (room?.t === 'e' && message?.file) {
-					Uploads.deleteOneById(message.file._id);
+					await Uploads.deleteOneById(message.file._id);
 				}
-				if (room?.t === 'e') { Messages.trashFindByIdAndRemove(id); }
+				if (room?.t === 'e') {
+					Messages.trashFindByIdAndRemove(id);
+				}
 				if (room?.t === 'e' && message) { broadcast('watch.messages', { clientAction, message }); }
 				break;
 			}
@@ -179,9 +181,12 @@ export function initWatchers(models: IModelsParam, broadcast: BroadcastCallback,
 			}
 
 			case 'removed': {
-				const trash = await Subscriptions.trashFindOneById<Pick<ISubscription, 'u' | 'rid'>>(id, { projection: { u: 1, rid: 1 } });
+				const trash = await Subscriptions.trashFindOneById<Pick<ISubscription, 'u' | 'rid' | 't'>>(id, { projection: { u: 1, rid: 1, t: 1 } });
 				const subscription = trash || { _id: id };
 				broadcast('watch.subscriptions', { clientAction, subscription });
+				if (trash?.t === 'e') {
+					Subscriptions.trashFindByIdAndRemove(id);
+				}
 				break;
 			}
 		}
@@ -325,8 +330,14 @@ export function initWatchers(models: IModelsParam, broadcast: BroadcastCallback,
 	});
 
 	watch<IRoom>(Rooms, async ({ clientAction, id, data, diff }) => {
+		console.log({ id });
 		if (clientAction === 'removed') {
-			Messages.deleteByRoomId(id);
+			const room = await Rooms.trashFindOneById<IRoom>(id, { projection: { t: 1 } });
+			if (room?.t === 'e') {
+				await Messages.deleteByRoomId(id);
+				Messages.trashRemoveByRoomId(id);
+				Rooms.trashFindByIdAndRemove(id);
+			}
 			broadcast('watch.rooms', { clientAction, room: { _id: id } });
 			return;
 		}
@@ -334,7 +345,6 @@ export function initWatchers(models: IModelsParam, broadcast: BroadcastCallback,
 		if (!hasRoomFields(data || diff)) {
 			return;
 		}
-
 		const room = data ?? await Rooms.findOneById(id, { projection: roomFields });
 		if (!room) {
 			return;
